feat(linter): split jest/prefer-each into separate vitest rule (#21967)
- part of #21664
