+++ crates/oxc_linter/src/rules/shared/jest_vitest/mod.rs
@@ -22,6 +22,7 @@ pub mod no_test_return_statement;
 pub mod no_unneeded_async_expect_function;
 pub mod prefer_called_with;
 pub mod prefer_comparison_matcher;
+pub mod prefer_each;
 pub mod prefer_expect_assertions;
 pub mod prefer_snapshot_hint;
 pub mod prefer_to_contain;
