+++ crates/oxc_linter/src/utils/mod.rs
@@ -37,8 +37,7 @@ pub use self::{
 // the crates/oxc_linter/data/vitest_compatible_jest_rules.json
 // file is also updated. The JSON file is used by the oxlint-migrate
 // and eslint-plugin-oxlint repos to keep everything synced.
-const VITEST_COMPATIBLE_JEST_RULES: [&str; 11] = [
-    "prefer-each",
+const VITEST_COMPATIBLE_JEST_RULES: [&str; 10] = [
     "prefer-equality-matcher",
     "prefer-expect-resolves",
     "prefer-hooks-in-order",
